refac tab switching components
