+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState } from "react";
+import { useRef, useState } from "react";
 import {
     FluentProvider,
     webDarkTheme,
@@ -11,6 +11,7 @@ import { ResultsWindow } from "./components/ResultsWindow";
 import { CustomEditor } from "./components/CustomEditor";
 import { ShortcutManager } from "./components/ShortcutManager";
 import { ModalDialog } from "./components/ModalDialog";
+import { TabSwitcher } from "./components/TabSwitcher";
 
 import { MenuBar } from "./components/MenuBar";
 import { ConnectionsMenu } from "./components/ConnectionsMenu";
@@ -28,27 +29,11 @@ type EditorTab = {
     results: Entity[];
 };
 
-type TabSwitchState = {
-    isCycling: boolean;
-    pressCount: number;
-    index: number;
-    show: boolean;
-    list: number[];
-};
-
 export default function App() {
     const [connectionsEnabled, setIsMenuOpen] = useState(true); 
     const [vimEnabled, setVimEnabled] = useState(true); 
     const [tabs, setTabs] = useState<EditorTab[]>([]);
     const [activeTabId, setActiveTabId] = useState(0);
-    const [tabMru, setTabMru] = useState<number[]>([]);
-    const [tabSwitch, setTabSwitch] = useState<TabSwitchState>({
-        isCycling: false,
-        pressCount: 0,
-        index: 0,
-        show: false,
-        list: [],
-    });
     const nextTabId = useRef(1);
     const [shortcutsOpen, setShortcutsOpen] = useState(false);
 
@@ -77,9 +62,6 @@ export default function App() {
     }
 
     const activeTab = tabs.find((tab) => tab.id === activeTabId);
-    const tabMap = new Map(tabs.map((tab) => [tab.id, tab]));
-    const tabSwitchList =
-        tabSwitch.isCycling && tabSwitch.list.length > 0 ? tabSwitch.list : tabMru;
 
     const handleAddTab = () => {
         const newId = nextTabId.current++;
@@ -118,81 +100,6 @@ export default function App() {
         onRetrieveResults(response);
     };
 
-    useEffect(() => {
-        const existingIds = new Set(tabs.map((tab) => tab.id));
-        setTabMru((prev) => {
-            const filtered = prev.filter((id) => existingIds.has(id));
-            const missing = tabs.map((tab) => tab.id).filter((id) => !filtered.includes(id));
-            return [...filtered, ...missing];
-        });
-    }, [tabs]);
-
-    useEffect(() => {
-        if (activeTabId === 0 || tabSwitch.isCycling) return;
-        setTabMru((prev) => {
-            const without = prev.filter((id) => id !== activeTabId);
-            return [activeTabId, ...without];
-        });
-    }, [activeTabId, tabSwitch.isCycling]);
-
-    useEffect(() => {
-        if (tabs.length < 2 && tabSwitch.isCycling) {
-            setTabSwitch({
-                isCycling: false,
-                pressCount: 0,
-                index: 0,
-                show: false,
-                list: [],
-            });
-        }
-    }, [tabs.length, tabSwitch.isCycling]);
-
-    useEffect(() => {
-        const handleKeyDown = (event: KeyboardEvent) => {
-            if (!event.ctrlKey || event.key !== "Tab") return;
-            if (tabMru.length < 2) return;
-
-            event.preventDefault();
-            event.stopPropagation();
-            event.stopImmediatePropagation();
-
-            setTabSwitch((prev) => {
-                const list = prev.isCycling ? prev.list : tabMru;
-                const nextIndex = prev.isCycling ? (prev.index + 1) % list.length : 1;
-                const nextPressCount = prev.isCycling ? prev.pressCount + 1 : 1;
-                const nextShow = nextPressCount > 1;
-                const nextId = list[nextIndex];
-                if (nextId) {
-                    setActiveTabId(nextId);
-                }
-                return {
-                    isCycling: true,
-                    pressCount: nextPressCount,
-                    index: nextIndex,
-                    show: nextShow,
-                    list,
-                };
-            });
-        };
-
-        const handleKeyUp = (event: KeyboardEvent) => {
-            if (event.key !== "Control") return;
-            setTabSwitch({
-                isCycling: false,
-                pressCount: 0,
-                index: 0,
-                show: false,
-                list: [],
-            });
-        };
-
-        window.addEventListener("keydown", handleKeyDown, true);
-        window.addEventListener("keyup", handleKeyUp, true);
-        return () => {
-            window.removeEventListener("keydown", handleKeyDown, true);
-            window.removeEventListener("keyup", handleKeyUp, true);
-        };
-    }, [tabMru]);
 
     return (
         <FluentProvider theme={webDarkTheme}>
@@ -229,27 +136,11 @@ export default function App() {
                         ))}
                     </div>
                 </ModalDialog>
-                {tabSwitch.show && tabSwitchList.length > 1 ? (
-                    <div className={styles.tabSwitcher} role="listbox" aria-label="Tabs">
-                        {tabSwitchList.map((id, index) => {
-                            const tab = tabMap.get(id);
-                            if (!tab) return null;
-                            const itemClasses = combineClasses(
-                                styles.tabSwitcherItem,
-                                index === tabSwitch.index && styles.tabSwitcherItemActive
-                            );
-                            return (
-                                <div
-                                    key={id}
-                                    className={itemClasses}
-                                    aria-selected={index === tabSwitch.index}
-                                >
-                                    <span className={styles.tabSwitcherTitle}>{tab.title}</span>
-                                </div>
-                            );
-                        })}
-                    </div>
-                ) : null}
+                <TabSwitcher
+                    tabs={tabs.map(({ id, title }) => ({ id, title }))}
+                    activeTabId={activeTabId}
+                    onTabSelect={setActiveTabId}
+                />
 
                 <div className={styles.wrapper}>
                     <ConnectionsMenu isOpen={connectionsEnabled} />
