+++ src/styles/AppStyles.tsx
@@ -138,35 +138,4 @@ export const useAppStyles = makeStyles({
         "&::-webkit-scrollbar-thumb:hover": { background: "#777777" },
         "&::-webkit-scrollbar-corner": { background: "#1f1f1f" },
     },
-    tabSwitcher: {
-        position: "fixed",
-        top: "72px",
-        left: "50%",
-        transform: "translateX(-50%)",
-        minWidth: "320px",
-        maxWidth: "520px",
-        backgroundColor: webDarkTheme.colorNeutralBackground2,
-        color: webDarkTheme.colorNeutralForeground1,
-        ...shorthands.border("1px", "solid", tokens.colorNeutralStroke2),
-        ...shorthands.borderRadius(tokens.borderRadiusMedium),
-        boxShadow: tokens.shadow16,
-        padding: tokens.spacingVerticalS,
-        zIndex: 1000,
-    },
-    tabSwitcherItem: {
-        display: "flex",
-        alignItems: "center",
-        gap: tokens.spacingHorizontalS,
-        padding: `${tokens.spacingVerticalXS} ${tokens.spacingHorizontalS}`,
-        borderRadius: tokens.borderRadiusSmall,
-    },
-    tabSwitcherItemActive: {
-        backgroundColor: tokens.colorNeutralBackground1Hover,
-    },
-    tabSwitcherTitle: {
-        overflow: "hidden",
-        textOverflow: "ellipsis",
-        whiteSpace: "nowrap",
-        flex: 1,
-    },
 });
